Integrating microservices

+++ limits-service/src/main/java/br/com/raphaelneves/microservices/limitsservice/config/LimitConfiguration.java
@@ -3,26 +3,28 @@ package br.com.raphaelneves.microservices.limitsservice.config;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 @ConfigurationProperties("limits-service")
 public class LimitConfiguration {
 
-    private int minimum;
-    private int maximum;
+    private BigDecimal minimum;
+    private BigDecimal maximum;
 
-    public int getMinimum() {
+    public BigDecimal getMinimum() {
         return minimum;
     }
 
-    public void setMinimum(int minimum) {
+    public void setMinimum(BigDecimal minimum) {
         this.minimum = minimum;
     }
 
-    public int getMaximum() {
+    public BigDecimal getMaximum() {
         return maximum;
     }
 
-    public void setMaximum(int maximum) {
+    public void setMaximum(BigDecimal maximum) {
         this.maximum = maximum;
     }
 }
